refactor(batch-selection): refactor component to TS
