--- src/components/batch-selection/index.d.ts
@@ -1,1 +1,0 @@
-export { default } from "./batch-selection";
