--- src/components/batch-selection/index.js
@@ -1,1 +1,0 @@
-export { default } from "./batch-selection.component";
