+++ src/components/batch-selection/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./batch-selection.component";
+export type { BatchSelectionProps } from "./batch-selection.component";
